+++ wn/__init__.py
@@ -9,6 +9,7 @@ __all__ = (
     'download',
     'add',
     'remove',
+    'export',
     'projects',
     'lexicons',
     'Lexicon',
@@ -30,6 +31,7 @@ from wn._exceptions import Error
 from wn._config import config  # noqa: F401
 from wn._db import is_schema_compatible
 from wn._add import add, remove
+from wn._export import export
 from wn._download import download
 from wn._core import (
     projects,